products db migrations

+++ database/migrations/2021_08_22_090843_create_users_permissions_table.php
@@ -14,8 +14,8 @@ class CreateUsersPermissionsTable extends Migration
     public function up()
     {
         Schema::create('users_permissions', function (Blueprint $table) {
-            $table->unsignedBigInteger('user_id');
-            $table->unsignedBigInteger('permission_id');
+            $table->unsignedInteger('user_id');
+            $table->unsignedInteger('permission_id');
 
             //FOREIGN KEY CONSTRAINTS
             $table->foreign('user_id')->references('id')->on('users');
