+++ database/migrations/2021_08_22_091011_create_roles_permissions_table.php
@@ -14,8 +14,8 @@ class CreateRolesPermissionsTable extends Migration
     public function up()
     {
         Schema::create('roles_permissions', function (Blueprint $table) {
-            $table->unsignedBigInteger('role_id');
-            $table->unsignedBigInteger('permission_id');
+            $table->unsignedInteger('role_id');
+            $table->unsignedInteger('permission_id');
 
             //FOREIGN KEY CONSTRAINTS
             $table->foreign('role_id')->references('id')->on('roles');
